Rename security.js to crypto.js
