+++ lib/crypto.js
