+++ metautil.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const utilities = require('./lib/utilities.js');
-const secutity = require('./lib/security.js');
+const crypto = require('./lib/crypto.js');
+const async = require('./lib/async.js');
 const semaphore = require('./lib/semaphore.js');
 const pool = require('./lib/pool.js');
-const async = require('./lib/async.js');
 
-module.exports = { ...utilities, ...secutity, ...semaphore, ...pool, ...async };
+module.exports = { ...utilities, ...crypto, ...semaphore, ...pool, ...async };
